Separate library source from example implementation
