+++ src/library.h
